still a wip, not sure how to make sure the content an shaders get copied wehn changed w/o rebuld all.. antialas didnt work  the shader clip mask was bleeding white.
work better...

added ARM target but works on droin in debug for now... not sure if it deploys  tested on the dx and gl desktop.. addded ios untests.

treed multisampling but doesnt work  in gl, also not sure w clipping how that should blend.. changed to nonpremultiped so mask can have hoes.. bascalicy if not whte then return transparent.   ..