disapble impllict usings.
set the versions to 303.. somehwo became wildcards.

retest touch fx.

in my other similar projec the embedded resource hack isn not needed... im not sure why.

the different i see is i use an extension called  MSBump on my Core lib... so mabye if this was added to the MGCore, it wouldnt not require the hack to embedded resource. i revisit this but i cant figure out if its the target folder paths or what.

adn if i touch an fx or resource it takes effect immendiately, and i dont  have to even do the embedded resource hack..... so i dont know aht this is the simplest way.    anywasy this project does work...now and the versions are correct...

the msbump is on github it changes teh core version every build.. this miight make the touch work .. i just dont have time to verfiyh..rightnow..and it has its own issues.
disable implicit usings and analyse make it simpler test droid dx , gl..

 {
  Configurations: {
    "Debug": {

      LabelDigits: 2
    },

    "Release": {
      BumpPatch: true,
      ResetLabel: "dev"
    }
  }
}