LOAD old 2 texture param clip shader that uses render target under #ifdef RENDERTARGETTESTcopy old 2 buffer render target sprite makinging blending clipper to a clip mask that just passes the sprite bacth trhoug one clip mask..
for now its just showing eithe rhte the mask or the sprite as a SANITY TEST

this is put in the Clipshader.. the original one that took 2 texture params was becauseI didnt know how to use registers.. its relatived to this issue
THE clip compare doesnt seem to wrk but it did on render target
this was tested on vs 2022 preview .net 7

https://community.monogame.net/t/passing-mask-texture-to-shader-works-in-xna-but-not-monogame/17342

the original file is ClipShaderSpriteTarget.fx

the #ifdef in com

on oepngl, windows dx , and androd ... all good.

the multismaple has to me zero on GL for rendertargets at least

this measn no antialias for lines to on gl.. its a konw defect.. when render to spirte its ok

but if we clip a large about of vector draws batched calls.. ( lines) or filled triangles.

its will show some pixellation..

STRATEGY behind all thsi... this is just basic test code.

hopoing to offload this to a community N platfrom project like
a shadertoy for monogame

wiht one shaded game core.. its easy to just test in secs 4 -5 platfroms , touch an fx , build and run.

NOTE teh compute branch is the futeure for monogame.

it not merged becasue of console issues. and IOS.

for macGL  linux GL windoow GL, windows DX, and android (GL) w xamarin
its ont a issuel

we ust no trimmiing and deploy as JIT, not AOT builds... this takes full advantage of SIMD
and the consoles are are not a target

appple stroe erquires vidoes and tons of monitoring even for a beta launch i hear
nothing but pain from devs on IOS..

they are turning to Metal.. or Vulcan or whatever and noone cares.. they say they will
drop open GL but then devs will leave them.

so for now we dont target iOS, only OSX.